Add archive course and improve error handling

+++ src/main/java/com/a10/mejabelajar/course/controller/CourseInformationController.java
@@ -1,8 +1,10 @@
 package com.a10.mejabelajar.course.controller;
 
+import com.a10.mejabelajar.auth.model.Teacher;
 import com.a10.mejabelajar.auth.model.User;
 import com.a10.mejabelajar.auth.service.TeacherService;
 import com.a10.mejabelajar.course.exception.CourseInformationInvalidException;
+import com.a10.mejabelajar.course.model.Course;
 import com.a10.mejabelajar.course.model.CourseInformation;
 import com.a10.mejabelajar.course.model.dto.CourseInformationDataTransferObject;
 import com.a10.mejabelajar.course.service.CourseInformationService;
@@ -41,14 +43,12 @@ public class CourseInformationController {
             @PathVariable int courseId,
             Model model) {
         var teacher = teacherService.getTeacherByUser(user);
-        var teacherCourse = teacher.getCourse();
-        if (teacherCourse == null) {
-            return REDIRECT_COURSE + teacher.getCourse().getId();
-        } else {
-            if (teacherCourse.getId() != courseId) {
-                return REDIRECT_COURSE + teacher.getCourse().getId();
-            }
+        var course = courseService.getCourseById(courseId);
+        String isValid = validateTeacherAccess(teacher, course, "Create Course Information");
+        if (!isValid.equals("")) {
+            return isValid;
         }
+
         model.addAttribute(COURSE_ID, courseId);
         model.addAttribute(COURSE_INFORMATION, new CourseInformationDataTransferObject());
         return "course/createCourseInformation";
@@ -65,20 +65,16 @@ public class CourseInformationController {
             Model model) {
 
         var teacher = teacherService.getTeacherByUser(user);
-        var teacherCourse = teacher.getCourse();
-        if (teacherCourse == null) {
-            return REDIRECT_COURSE + teacher.getCourse().getId();
-        } else {
-            if (teacherCourse.getId() != courseId) {
-                return REDIRECT_COURSE + teacher.getCourse().getId();
-            }
+        var course = courseService.getCourseById(courseId);
+        String isValid = validateTeacherAccess(teacher, course, "Create Course Information");
+        if (!isValid.equals("")) {
+            return isValid;
         }
 
         try {
             var courseInformation = new CourseInformation();
             modelMapper.map(courseInformationDataTransferObject, courseInformation);
 
-            var course = courseService.getCourseById(courseId);
             courseInformation.setCourse(course);
             courseInformationService.createCourseInformation(courseInformation);
             return REDIRECT_COURSE + courseId;
@@ -100,13 +96,10 @@ public class CourseInformationController {
             @PathVariable int courseInformationId,
             Model model) {
         var teacher = teacherService.getTeacherByUser(user);
-        var teacherCourse = teacher.getCourse();
-        if (teacherCourse == null) {
-            return REDIRECT_COURSE + teacher.getCourse().getId();
-        } else {
-            if (teacherCourse.getId() != courseId) {
-                return REDIRECT_COURSE + teacher.getCourse().getId();
-            }
+        var course = courseService.getCourseById(courseId);
+        String isValid = validateTeacherAccess(teacher, course, "Update Course Information");
+        if (!isValid.equals("")) {
+            return isValid;
         }
 
         var courseInformation =
@@ -129,13 +122,10 @@ public class CourseInformationController {
             Model model) {
 
         var teacher = teacherService.getTeacherByUser(user);
-        var teacherCourse = teacher.getCourse();
-        if (teacherCourse == null) {
-            return REDIRECT_COURSE + teacher.getCourse().getId();
-        } else {
-            if (teacherCourse.getId() != courseId) {
-                return REDIRECT_COURSE + teacher.getCourse().getId();
-            }
+        var course = courseService.getCourseById(courseId);
+        String isValid = validateTeacherAccess(teacher, course, "Update Course Information");
+        if (!isValid.equals("")) {
+            return isValid;
         }
 
         try {
@@ -159,21 +149,38 @@ public class CourseInformationController {
     /**
      * Delete course information for specific course specified by id in path variable.
      */
-    @GetMapping(value = "/delete/{courseId}/{courseInformationId}")
+    @PostMapping(value = "/delete/{courseId}/{courseInformationId}")
     public String deleteCourseInformation(
             @AuthenticationPrincipal User user,
             @PathVariable int courseId,
             @PathVariable int courseInformationId) {
         var teacher = teacherService.getTeacherByUser(user);
-        var teacherCourse = teacher.getCourse();
-        if (teacherCourse == null) {
-            return REDIRECT_COURSE + teacher.getCourse().getId();
-        } else {
-            if (teacherCourse.getId() != courseId) {
-                return REDIRECT_COURSE + teacher.getCourse().getId();
-            }
+        var course = courseService.getCourseById(courseId);
+        String isValid = validateTeacherAccess(teacher, course, "Delete Course Information");
+        if (!isValid.equals("")) {
+            return isValid;
         }
         courseInformationService.deleteCourseInformationById(courseInformationId);
         return REDIRECT_COURSE + courseId;
+    }
+
+    /**
+     * Validate teacher access to a course information.
+     */
+    public String validateTeacherAccess(Teacher teacher, Course course, String action) {
+        if (course.isArchived()) {
+            if (course.getTeacher() == teacher) {
+                return REDIRECT_COURSE + course.getId() + "?error=This course is archived";
+            }
+        }
+        if (course.getTeacher() != teacher) {
+            if (!teacher.isHaveCourse()) {
+                return REDIRECT_COURSE + "?error=You don't have access to " + action;
+            }
+            return REDIRECT_COURSE
+                    + courseService.getCourseByTeacherAndStatus(teacher, false).getId()
+                    + "?error=You don't have access to " + action;
+        }
+        return "";
     }
 }
